Major Bug Fixes

+++ botasaurus_driver/core/browser.py
@@ -1,22 +1,19 @@
 from __future__ import annotations
 
-import asyncio
+import subprocess
 import atexit
-import json
-import urllib.parse
-import urllib.request
 import os
-from typing import List, Union, Tuple
+from typing import List, Union
 
+import time
+import requests
 from .profiles import delete_profile, get_target_folders, run_check_and_delete_in_thread
-from ..exceptions import DriverException
 from .. import cdp
 from . import util
 from . import tab
-from ._contradict import ContraDict
 from .config import PathLike, Config, free_port, is_posix
 from .connection import Connection
-from .custom_storage_cdp import  get_cookies, set_cookies
+from .custom_storage_cdp import get_cookies, set_cookies
 
 import os
 import signal
@@ -36,17 +33,43 @@ def get_folder_name_from_path(absolute_path):
     """
     return os.path.basename(absolute_path)
 
+def ensure_chrome_is_alive(url):
+    retries = 10
+    delay = 0.25
+    timeout = 5  # Adjust this value based on your requirements
+
+    for _ in range(retries):
+        try:
+            response = requests.get(url, timeout=timeout)
+            response.raise_for_status()
+            return response.json()
+        except requests.exceptions.RequestException:
+            # print(f"Attempt {i+1} failed: {str(e)}")
+            time.sleep(delay)
+
+    raise Exception(f"Failed to connect to Chrome URL: {url}.")
+
+def wait_for_graceful_close(_process: subprocess.Popen):
+    retries = 10
+    delay = 0.05
+
+    for _ in range(retries):
+        if _process.poll() is not None:
+            return True
+        time.sleep(delay)
+
+    return _process.poll() is not None
+
 class Browser:
-    _process: asyncio.subprocess.Process
+    _process: subprocess.Popen
     _process_pid: int
-    _http: HTTPApi = None
     _cookies: CookieJar = None
 
     config: Config
     connection: Connection
 
     @classmethod
-    async def create(
+    def create(
         cls,
         config: Config = None,
         *,
@@ -70,24 +93,16 @@ class Browser:
                 **kwargs,
             )
         instance = cls(config)
-        await instance.start()
+        instance.start()
         return instance
 
-    def __init__(self, config: Config, **kwargs):
+    def __init__(self, config: Config):
         """
         constructor. to create a instance, use :py:meth:`Browser.create(...)`
 
         :param config:
         """
 
-        try:
-            asyncio.get_running_loop()
-        except RuntimeError:
-            raise RuntimeError(
-                "{0} objects of this class are created using await {0}.create()".format(
-                    self.__class__.__name__
-                )
-            )
         # weakref.finalize(self, self._quit, self)
         self.config = config
 
@@ -98,12 +113,12 @@ class Browser:
         self._process = None
         self._process_pid = None
         self._keep_profile_directory = None
-        self._is_updating = asyncio.Event()
+        self._is_updating = False
         self.connection: Connection = None
 
     @property
     def websocket_url(self):
-        return self.info.webSocketDebuggerUrl
+        return self.info['webSocketDebuggerUrl']
 
     @property
     def main_tab(self) -> tab.Tab:
@@ -150,8 +165,8 @@ class Browser:
                     lambda item: item.target_id == target_info.target_id, self.targets
                 )
             )
-            current_target = current_tab.target
-
+            # todo: maybe connections need to be reinited?
+            current_tab.target = target_info
 
         elif isinstance(event, cdp.target.TargetCreated):
             target_info: cdp.target.TargetInfo = event.target_info
@@ -161,21 +176,21 @@ class Browser:
                 (
                     f"ws://{self.config.host}:{self.config.port}"
                     f"/devtools/page"  # all types are 'page' internally in chrome apparently
-                    f"/{target_info.target_id}"                ),
+                    f"/{target_info.target_id}"
+                ),
                 target=target_info,
                 browser=self,
             )
             self.targets.append(new_target)
 
-
         elif isinstance(event, cdp.target.TargetDestroyed):
             current_tab = next(
                 filter(lambda item: item.target_id == event.target_id, self.targets)
             )
-            
+            current_tab.close_connections()
             self.targets.remove(current_tab)
 
-    async def get(
+    def get(
         self, url="chrome://welcome", new_tab: bool = False, new_window: bool = False, referrer=None
     ) -> tab.Tab:
         """top level get. utilizes the first tab to retrieve given url.
@@ -191,7 +206,7 @@ class Browser:
         """
         if new_tab or new_window:
             # creat new target using the browser session
-            target_id = await self.connection.send(
+            target_id = self.connection.send(
                 cdp.target.create_target(
                     url, new_window=new_window, enable_begin_frame_control=True
                 )
@@ -208,116 +223,71 @@ class Browser:
             # first tab from browser.tabs
             connection = next(filter(lambda item: item.type_ == "page", self.targets))
             # use the tab to navigate to new url
-            frame_id, loader_id, *_ = await connection.send(cdp.page.navigate(url, referrer=referrer))
+            frame_id, _, *_ = connection.send(cdp.page.navigate(url, referrer=referrer))
             # update the frame_id on the tab
             connection.frame_id = frame_id
 
-        await connection.sleep(0.25)
+        time.sleep(0.25)
         return connection
 
-    async def start(self=None) -> Browser:
-        """launches the actual browser"""
-        if not self:
-            print("use ``await Browser.create()`` to create a new instance")
-            return
-        if self._process or self._process_pid:
-            if self._process.returncode is not None:
-                return await self.create(config=self.config)
-            print("ignored! this call has no effect when already running.")
-            return
+    def start(self=None) -> Browser:
 
         self.config.host = self.config.host or "127.0.0.1"
         self.config.port = self.config.port or free_port()
         exe = self.config.browser_executable_path
         params = self.config()
-        self._process: asyncio.subprocess.Process = (
-            await asyncio.create_subprocess_exec(
-                # self.config.browser_executable_path,
-                # *cmdparams,
-                exe,
-                *params,
-                stdin=asyncio.subprocess.PIPE,
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-                close_fds=is_posix,
-            )
+        chrome_process = subprocess.Popen(
+            [exe, *params],
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+            close_fds=is_posix,
         )
+
+        self._process = chrome_process
 
         self._process_pid = self._process.pid
 
-        self._http = HTTPApi((self.config.host, self.config.port))
         self.base_folder_name = get_folder_name_from_path(self.config.profile_directory)
         instances = util.get_registered_instances()
         instances.add(self)
-        
-        await asyncio.sleep(0.25)
-        for _ in range(5):
-            try:
-                self.info = ContraDict(await self._http.get("version"), silent=True)
-            except (Exception,):
-                if _ == 4:
-                    pass
-                await asyncio.sleep(0.5)
-            else:
-                break
+        chrome_url = f"http://{self.config.host}:{self.config.port}/json/version"
 
-        if not self.info:
-            raise DriverException(
-                (
-                    """
-                ---------------------
-                Failed to connect to browser
-                ---------------------
-                One of the causes could be when you are running as root.
-                """
-                )
-            )
+        self.info = ensure_chrome_is_alive(chrome_url)
 
-        self.connection = Connection(self.info.webSocketDebuggerUrl, _owner=self)
+        self.connection = Connection(self.info['webSocketDebuggerUrl'], _owner=self)
 
-        if self.config.autodiscover_targets:
+        self.connection.handlers[cdp.target.TargetInfoChanged] = [
+            self._handle_target_update
+        ]
+        self.connection.handlers[cdp.target.TargetCreated] = [
+            self._handle_target_update
+        ]
+        self.connection.handlers[cdp.target.TargetDestroyed] = [
+            self._handle_target_update
+        ]
+        self.connection.handlers[cdp.target.TargetCrashed] = [
+            self._handle_target_update
+        ]
+        self.connection.send(cdp.target.set_discover_targets(discover=True))
+        # self.connection.wait_to_be_idle()
+        self.update_targets()
+        # await self
 
-            # self.connection.add_handler(
-            #     cdp.target.TargetInfoChanged, self._handle_target_update
-            # )
-            # self.connection.add_handler(
-            #     cdp.target.TargetCreated, self._handle_target_update
-            # )
-            # self.connection.add_handler(
-            #     cdp.target.TargetDestroyed, self._handle_target_update
-            # )
-            # self.connection.add_handler(
-            #     cdp.target.TargetCreated, self._handle_target_update
-            # )
-            #
-            self.connection.handlers[cdp.target.TargetInfoChanged] = [
-                self._handle_target_update
-            ]
-            self.connection.handlers[cdp.target.TargetCreated] = [
-                self._handle_target_update
-            ]
-            self.connection.handlers[cdp.target.TargetDestroyed] = [
-                self._handle_target_update
-            ]
-            self.connection.handlers[cdp.target.TargetCrashed] = [
-                self._handle_target_update
-            ]
-            await self.connection.send(cdp.target.set_discover_targets(discover=True))
-        await self
+        # CLEAN UO
         fls = get_target_folders(instances)
-        
+
         if fls:
             run_check_and_delete_in_thread(fls)
         # self.connection.handlers[cdp.inspector.Detached] = [self.close]
         # return self
 
-    async def _get_targets(self) -> List[cdp.target.TargetInfo]:
-        info = await self.connection.send(cdp.target.get_targets(), _is_update=True)
+    def _get_targets(self) -> List[cdp.target.TargetInfo]:
+        info = self.connection.send(cdp.target.get_targets(), _is_update=True)
         return info
 
-    async def update_targets(self):
+    def update_targets(self):
         targets: List[cdp.target.TargetInfo]
-        targets = await self._get_targets()
+        targets = self._get_targets()
         for t in targets:
             for existing_tab in self.targets:
                 existing_target = existing_tab.target
@@ -325,26 +295,25 @@ class Browser:
                     existing_tab.target.__dict__.update(t.__dict__)
                     break
             else:
-
+                # new target
+                print('Making a New Connection')
                 self.targets.append(
                     Connection(
                         (
                             f"ws://{self.config.host}:{self.config.port}"
                             f"/devtools/page"  # all types are 'page' somehow
                             f"/{t.target_id}"
-                                                                            ),
+                        ),
                         target=t,
                         _owner=self,
                     )
                 )
+        time.sleep(0)
 
-        await asyncio.sleep(0)
-
-    async def __aenter__(self):
-
+    def __enter__(self):
         return self
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(self, exc_type, exc_val, _):
         if exc_type and exc_val:
             raise exc_type(exc_val)
 
@@ -368,69 +337,67 @@ class Browser:
                 else:
                     del self._i
 
-
     def close(self):
-        # No need we are killing
-        # try:
-        #     # asyncio.get_running_loop().create_task(self.connection.send(cdp.browser.close()))
-        #     asyncio.get_event_loop().create_task(self.connection.aclose())
-        # except RuntimeError:
-        #     if self.connection:
-        #         try:
-        #             # asyncio.run(self.connection.send(cdp.browser.close()))
-        #             asyncio.run(self.connection.aclose())
-        #         except Exception:
-        #             pass
-        # except Exception:   
-        #     pass
-        
-        for _ in range(3):
+        # close gracefully
+        self.close_chrome()
+        self.close_tab_connections()
+        self.close_browser_connection()
+        if not wait_for_graceful_close(self._process):
+        #   process not closed
             try:
-                # print("killing browser")
                 self._process.terminate()
-
-                # print("killed browser")            
-                break
+                self._process.wait()
             except (Exception,):
                 try:
                     self._process.kill()
-                    break
                 except (Exception,):
                     try:
-                        if hasattr(self, "browser_process_pid"):
-                            kill_process(self._process_pid)
-                            break
+                        kill_process(self._process_pid)
                     except (TypeError,):
                         pass
                     except (PermissionError,):
-                        
                         pass
                     except (ProcessLookupError,):
                         pass
                     except (Exception,):
                         raise
-            self._process = None
-            self._process_pid = None
+        self._process = None
+        self._process_pid = None
 
-        
         if self.config.is_temporary_profile:
             delete_profile(self.config.profile_directory)
         self.config.close()
-        
-    def __await__(self):
-        # return ( asyncio.sleep(0)).__await__()
-        return self.update_targets().__await__()
+
+    def close_browser_connection(self):
+        try:
+          self.connection.close()
+        except Exception as e:
+            print(e)
+
+    def close_tab_connections(self):
+        try:
+            while self.targets:
+                # close just connections, chrome tabs are closed 
+                self.targets.pop().close_connections()
+        except Exception as e:
+            print(e)
+
+    def close_chrome(self):
+        try:
+          self.connection.send(cdp.browser.close())
+        except Exception as e:
+            print(e)
+
 
     def __del__(self):
         pass
-
 
 class CookieJar:
     def __init__(self, browser: Browser):
         self._browser = browser
         # self._connection = connection
 
-    async def get_all(
+    def get_all(
         self, requests_cookie_format: bool = False
     ) -> List[Union[cdp.network.Cookie]]:
         """
@@ -450,7 +417,7 @@ class CookieJar:
             break
         else:
             connection = self._browser.connection
-        cookies = await connection.send(get_cookies())
+        cookies = connection.send(get_cookies())
         if requests_cookie_format:
             import requests.cookies
 
@@ -467,7 +434,7 @@ class CookieJar:
             ]
         return cookies
 
-    async def set_all(self, cookies: List[cdp.network.CookieParam]):
+    def set_all(self, cookies: List[cdp.network.CookieParam]):
         """
         set cookies
 
@@ -484,8 +451,9 @@ class CookieJar:
             break
         else:
             connection = self._browser.connection
-        await connection.send(set_cookies(cookies))
-    async def clear(self):
+        connection.send(set_cookies(cookies))
+
+    def clear(self):
         """
         clear current cookies
 
@@ -502,35 +470,6 @@ class CookieJar:
             break
         else:
             connection = self._browser.connection
-        await connection.send(cdp.storage.clear_cookies())
-
-
-class HTTPApi:
-    def __init__(self, addr: Tuple[str, int]):
-        self.host, self.port = addr
-        self.api = "http://%s:%d" % (self.host, self.port)
-
-    async def get(self, endpoint: str):
-        return await self._request(endpoint)
-
-    async def _request(self, endpoint, method: str = "get", data: dict = None):
-        url = urllib.parse.urljoin(
-            self.api, f"json/{endpoint}" if endpoint else "/json"
-        )
-        if data and method.lower() == "get":
-            raise DriverException("get requests cannot contain data")
-        if not url:
-            url = self.api + endpoint
-        request = urllib.request.Request(url)
-        request.method = method
-        request.data = None
-        if data:
-            request.data = json.dumps(data).encode("utf-8")
-
-        response = await asyncio.get_running_loop().run_in_executor(
-            None, urllib.request.urlopen, request
-        )
-        return json.loads(response.read())
-
+        connection.send(cdp.storage.clear_cookies())
 
 atexit.register(util.deconstruct_browser)
